src/python/requirements.txt
# SHA1:3102645e724938384fee7ede7865bc3f0fe12809
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
aws-secretsmanager-caching==1.1.1.5
    # via hmd-cli-tools
boto3==1.21.7
    # via hmd-cli-tools
botocore==1.24.9
    # via
    #   aws-secretsmanager-caching
    #   boto3
    #   hmd-cli-tools
    #   s3transfer
cement==3.0.6
    # via hmd-cli-tools
certifi==2021.10.8
    # via
    #   hmd-cli-tools
    #   requests
cffi==1.15.1
    # via pynacl
charset-normalizer==2.0.12
    # via
    #   hmd-cli-tools
    #   requests
deprecated==1.2.13
    # via pygithub
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via -r C:\Users\KateWalls\programs\hmd-tf-bartleby\src\python\requirements.in
hmd-cli-tools~=0.2.173
    # via -r C:\Users\KateWalls\programs\hmd-tf-bartleby\src\python\requirements.in
idna==3.3
    # via
    #   hmd-cli-tools
    #   requests
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
    #   hmd-cli-tools
pycparser==2.21
    # via cffi
pygithub==1.55
    # via -r C:\Users\KateWalls\programs\hmd-tf-bartleby\src\python\requirements.in
pyjwt==2.4.0
    # via pygithub
pynacl==1.5.0
    # via pygithub
python-dateutil==2.8.2
    # via
    #   botocore
    #   hmd-cli-tools
python-dotenv==0.19.2
    # via hmd-cli-tools
requests==2.27.1
    # via
    #   hmd-cli-tools
    #   pygithub
s3transfer==0.5.2
    # via
    #   boto3
    #   hmd-cli-tools
six==1.16.0
    # via
    #   hmd-cli-tools
    #   python-dateutil
smmap==5.0.0
    # via gitdb
urllib3==1.26.8
    # via
    #   botocore
    #   hmd-cli-tools
    #   requests
wrapt==1.14.1
    # via deprecated
